Use fixed width type (u32) instead of size_t for face indices

+++ examples/wuzy-gjk-debug.cpp
@@ -68,7 +68,7 @@ int main()
          1.0f,  1.0f, -1.0f,
         // clang-format on
     };
-    const std::vector<size_t> triangle_indices { 0, 1, 2 };
+    const std::vector<uint32_t> triangle_indices { 0, 1, 2 };
 
     wuzy::ConvexPolyhedronCollider collider_a(triangle_positions, triangle_indices);
